requirements.txt
chardet==5.2.0
numpy==2.2.3
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
qrcode==8.0
reportlab==4.3.1
six==1.17.0
tzdata==2025.1
